Added a new light engine and the back work to coloured lighting.

+++ src/main/java/tgw/evolution/mixin/MixinItemStack.java
@@ -117,7 +117,7 @@ public abstract class MixinItemStack implements PatchItemStack {
     public abstract Item getItem();
 
     @Override
-    public byte getLightEmission() {
+    public short getLightEmission() {
         if (this.isEmpty()) {
             return 0;
         }
